pyproject.toml
[tool.poetry]
name = "zendesk"
version = "1.4.7"
description = "Zendesk Book allows users to manage Zendesk tickets and users."
authors = ["Emiliano Lesende <emiliano.lesende@gmail.com>"]
readme = "README.md"
packages = [{include = "zendesk", from="src"}]

[cloud]
AWS_REGION = "us-west-2"
AWS_ACCOUNT_ID = "719468614044"
DOCKER_REGISTRY = "{AWS_ACCOUNT_ID}.dkr.ecr.{AWS_REGION}.amazonaws.com"
TERRAFORM_BACKEND_BUCKET_NAME = "bdk-terraform-backend"

[environment]
bdk_runtime_version = "1.25.5"
bdk_image_repo = "719468614044.dkr.ecr.us-west-2.amazonaws.com/kognitos/bdk"


[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
requests = "^2.31.0"
kognitos-bdk-api = "^1.2"
kognitos-bdk-klang = "^2.2"
opentelemetry-api = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-exporter-otlp = "^1.26.0"
opentelemetry-instrumentation = "^0.47b0"
opentelemetry-instrumentation-requests = "^0.47b0"
zenpy = "^2.0.56"
boto3 = "^1.36.14"
pytest = "^8.3.4"
autopep8 = "^2.3.2"
dotenv = "^0.9.9"
asyncio = "^3.4.3"
pytest-asyncio = "^0.25.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^24.2.0"
pylint = "^3.0.3"
pyright = "^1.1.360"
pre-commit = "^3.7.1"
boto3 = "^1.34.139"
pytest-recording = "^0.13.1"
kognitos-bdk-linter = "^1.6"

[tool.poetry.scripts]
tests = "poetry_scripts:run_tests"
format = "poetry_scripts:run_format"
lint = "poetry_scripts:run_lint"
record = "poetry_scripts:run_record"
type-check = "poetry_scripts:run_type_check"
doc = "poetry_scripts:run_doc"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
include = ["data/*"]

[tool.pylint.MAIN]
py-version = "3.11"
load-plugins = ["kognitos.bdk.linter"]

[tool.pylint.messages_control]
disable = [
    "line-too-long",
    "redefined-outer-name",
    "too-many-instance-attributes",
    "unused-import",
    "duplicate-code",
    "protected-access"
]

[tool.poetry.plugins."kognitos-book"]
zendesk = "zendesk.book:ZendeskBook"
